refactor(gui): move ColorPalette to debug subfolder
- Create components/debug/ for debug-only components
- Move ColorPalette.tsx to debug/ColorPalette.tsx
- Fix import paths: ./ui/* → ../ui/*
- Update import in App.jsx
