+++ cmd/gui/frontend/src/components/debug/ColorPalette.tsx
@@ -1,5 +1,5 @@
-import { Button } from "./ui/button";
-import { Card } from "./ui/card";
+import { Button } from "../ui/button";
+import { Card } from "../ui/card";
 import { ArrowLeft } from "lucide-react";
 
 interface ColorPaletteProps {
